+++ Mark/DiagramDownApp.swift
@@ -14,8 +14,11 @@ struct DiagramDownApp: App {
 
     var body: some Scene {
         DocumentGroup(newDocument: MarkdownDocument()) { file in
-            ContentView(document: file.$document)
+            ContentView(document: file.$document, fileURL: file.fileURL)
                 .preferredColorScheme(preferredColorScheme)
+        }
+        .commands {
+            PreviewCommands()
         }
 
         Settings {